requirements.txt
asgiref==3.5.2
autopep8==1.5.4
backports.zoneinfo==0.2.1
cachetools==4.1.1
certifi==2020.4.5.1
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.0.12
cloudinary==1.29.0
coreapi==2.3.3
coreschema==0.0.4
coverage==6.4.2
cryptography==37.0.4
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2
django-cloudinary-storage==0.3.0
django-cors-headers==3.5.0
django-filter==22.1
django-rest-auth==0.9.5
django-rest-passwordreset==1.2.1
django-storages==1.12.3
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-action-serializer==1.0.5
drf-writable-nested==0.6.4
drf-yasg==1.20.0
dropbox==11.33.0
facebook-sdk==3.1.0
Faker==4.1.0
future==0.18.2
google-api-core==1.22.2
google-api-python-client==1.12.1
google-auth==1.21.2
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.9
inflection==0.4.0
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
oauthlib==3.1.0
packaging==20.4
Pillow==9.2.0
ply==3.11
protobuf==3.13.0
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.21
PyJWT==2.4.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.6
python-twitter==3.5
python3-openid==3.2.0
pytz==2020.1
requests==2.28.1
requests-oauthlib==1.3.0
rsa==4.6
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.3.1
stone==3.3.1
text-unidecode==1.3
toml==0.10.1
uritemplate==3.0.1
urllib3==1.26.10
whitenoise==6.0.0
xlwt==1.3.0
